% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_commented.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the results on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer with the state id}

\item{year}{An integer with the year to use}
}
\description{
It plots the results on a map from a given year and state with the fatal crasesh.
}
\examples{
\dontrun{fars_map_state(12, 2012)}

}
