% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_commented.R
\name{fars_read}
\alias{fars_read}
\title{Loads a CSV file with FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string containing the path that it is going to be used for reading the data.}
}
\value{
Returns a data.frame containing the data.
}
\description{
It checks if the CSV file exists. 
If it exists, it reads the data, if it does not, it will return an error.
}
\examples{
\dontrun{
FARSdf<-fars_read('data.csv.bz2')
}

}
