% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_commented.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Processes data from the wanted years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer vector with the year(s) to use.}
}
\value{
A list of data.frames, each data.frame (tibble in this case) contains two columns month and year and a row for each fatal accident.
}
\description{
It processes data for a given year to obtain the crashes from each month. If the years selected does not exist it returns an error.
}
\examples{
\dontrun{fry <- fars_read_years(2011:2015)}

}
