% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_commented.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer vector of years where it will search}
}
\value{
A data.frame with two columns, year and month. Each row the sum of observations in a year-month.
}
\description{
It summarize the number of observations by month and year in the given year(s)
}
\examples{
\dontrun{fsy<-fars_summarize_years(2011:2015)}

}
