% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_commented.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer where the year is stored.}
}
\value{
Returns a string with the filename
}
\description{
It creates a filename taking into account the year.
}
\examples{
\dontrun{fn<-make_filename(2017)}

}
